refactor: extract useEstimatePricing + form-helpers (Phase 1a)
Move shared LineItemRow/OptionTier types and pure helpers (parseCents,
lineTotal, mapPrepLevel, constants) to lib/estimates/form-helpers.ts.
Extract all pricing computations (paintingResult, totals, reviewTotal)
into useEstimatePricing hook. useEstimateForm drops from 1046 → 924
lines with zero behavior change.
